refactor: 멀티 모듈 패키지 구조 리팩토링 (#34)
* refactor: 테스트 환경변수 통일

* refactor: config 패키지 domain 모듈로 이동

* refactor: 패키지 구조 변경

* refactor: MinIoTest local qr 코드 저장 코드 삭제

* refactor: google domain 패키지 구조 변경

* fix: generate qr code

* refactor: dto 패키지 위치 변경

* refactor: dto 패키지 위치 변경
