--- qrust-common/src/main/java/com/qrust/config/SwaggerConfig.java
@@ -1,27 +1,0 @@
-package com.qrust.config;
-
-
-import io.swagger.v3.oas.models.info.Info;
-import org.springdoc.core.models.GroupedOpenApi;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-
-    @Bean
-    public GroupedOpenApi qrustGroupedOpenApi() {
-        return GroupedOpenApi
-                .builder()
-                .group("capstone")
-                .pathsToMatch("/**")
-                .addOpenApiCustomizer(openApi ->
-                        openApi.setInfo(new Info()
-                                .title("qrust api")
-                                .description("QRust API")
-                                .version("1.0.0")
-                        )
-                )
-                .build();
-    }
-}
