--- qrust-common/src/main/java/com/qrust/infrastructure/google/config/GoogleSafeBrowsingConfig.java
@@ -1,15 +1,0 @@
-package com.qrust.infrastructure.google.config;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class GoogleSafeBrowsingConfig {
-
-    @Value("${google.safe-browsing.api.key}")
-    private String apiKey;
-
-    public String getApiKey() {
-        return apiKey;
-    }
-}
