+++ qrust-domain/src/main/java/com/qrust/external/google/application/dto/response/GoogleSafeBrowsingResponse.java
