+++ qrust-domain/src/main/java/com/qrust/qrcode/dto/QrCodeGenerateRequestDto.java
